requirements.txt
streamlit==1.32.2
python-dotenv==1.0.1
openai==1.30.1

boto3